Improved OBJ loader and added UVSet class
Added error detection to OBJ loader to throw an error if a pointer
points to a UV or vertex that has not been defined, will also log if a
UV for a vertex was overridden.
Created UVSet class as was previously using Vertex3 with no z usage
which was a waste of memory.